diplom_back/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cached-property==2.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
networkx==3.4.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMuPDF==1.25.5
PyPDF2==3.0.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlalchemy-orm==1.2.10
starlette==0.46.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
